fix(interest): box sizing and legal disclaimer placement
